Final version

+++ delete.php
@@ -11,7 +11,7 @@
 
 
     $id = $_GET['id'];
-    $statement = $conn->prepare("SELECT * from contacts where id=:id");
+    $statement = $conn->prepare("SELECT * from contacts where id=:id limit 1");
     $statement->execute([":id" => $id]);
 
     if(!$statement){
@@ -24,6 +24,18 @@
     // $statement->bindParam(":id", $id);
     // $statement->execute();
 
+    $contact = $statement->fetch(PDO::FETCH_ASSOC);
+
+    if($contact["user_id"] !== $_SESSION["user"]['id']){
+        http_response_code(403);
+        print("HTTP 403 - Unauthorized");
+        return;
+    }
+
     $conn->prepare("DELETE FROM contacts where id=:id")->execute([":id" => $id]);
+
+    $_SESSION["flash"] = ["message" => "Contact {$contact['name']} deleted successfully."];
+
     header("Location: home.php");
+    return;
 ?>
